# base/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Học Tiếng Anh",
        "description": "Tài liệu API cho ứng dụng học tiếng Anh.",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/lesson/list": {
            "get": {
                "tags": [
                    "Lesson"
                ],
                "summary": "Lấy danh sách bài học",
                "description": "Trả về danh sách các bài học, có thể lọc theo category_id",
                "operationId": "2eb38a6bded42f80356c7aa746fd7d07",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "ID danh mục cần lọc",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Trang hiện tại",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách bài học",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Bài học đầu tiên"
                                                    },
                                                    "content": {
                                                        "type": "string",
                                                        "example": "Nội dung bài học..."
                                                    },
                                                    "category_id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "total_pages": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "total_items": {
                                            "type": "integer",
                                            "example": 50
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/lessons/{id}": {
            "get": {
                "tags": [
                    "Lesson"
                ],
                "summary": "Lấy chi tiết bài học theo ID",
                "description": "Trả về chi tiết một bài học cụ thể",
                "operationId": "f63f80bce34ecb1b23418b41ee20d495",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID bài học",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thông tin chi tiết bài học",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Bài học đầu tiên"
                                        },
                                        "content": {
                                            "type": "string",
                                            "example": "Nội dung chi tiết bài học..."
                                        },
                                        "category_id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy bài học"
                    }
                }
            }
        },
        "/api/v1/lesson/choice/{id}": {
            "get": {
                "tags": [
                    "Lesson"
                ],
                "summary": "Lấy một bài học",
                "operationId": "d02b21d9e0f74dc22c97db543f2955a3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bài học",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "I work as a software developer."
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Tôi làm việc như một lập trình viên phần mềm."
                                        },
                                        "tusapxep": {
                                            "type": "string",
                                            "example": "lập, phần, như, tôi, một, làm, viên, phần, developer, work"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}